GenEntityKey -> Internal.GenerateKeyInternal, expose GenerateKey subclass
Suppose you have a user doing something like this:

  meep
    :: forall a m.
      (Dependencies a ~ (), GraphulaContext m '[a], GenerateKey a)
      => m (Entity a)
  meep = node () mempty

e.g.

  x <- meep @A

If we define

  type GenerateKey a = (GenerateKeyInternal (KeySource a) a, KeyConstraint (KeySource a) a)

then GHC looks through GenerateKey and spits out

    • Non type-variable argument
        in the constraint: Graphula.Internal.GenerateKeyInternal
                             (KeySource a) a
      (Use FlexibleContexts to permit this)

Which isn't terrible - the user can just turn on FlexibleContexts.
Unfortunately, the user is also looking at stuff they don't have access
to (without importing Graphula.Internal). Instead we can define this
subclass for every type:

  class (GenerateKeyInternal (KeySource a) a, KeyConstraint (KeySource a) a) => GenerateKey a
  instance (GenerateKeyInternal (KeySource a) a, KeyConstraint (KeySource a) a) => GenerateKey a

Now `meep` just works since we're handling the complicated constraints ourselves.
